Изменен тип данных у id

+++ src/main/java/mem/MrPonerYea/PoolControl/exception/EntityDoesNotExistException.java
@@ -3,7 +3,7 @@ package mem.MrPonerYea.PoolControl.exception;
 import org.springframework.http.HttpStatus;
 
 public class EntityDoesNotExistException extends BaseException {
-    public EntityDoesNotExistException(Integer id, Class<?> entityClass) {
+    public EntityDoesNotExistException(Long id, Class<?> entityClass) {
         super(ErrorCode.ENTITY_DOES_NOT_EXIST, HttpStatus.NOT_FOUND,
                 "Объект " + entityClass.getSimpleName() + " с айди " + id.toString() + " не существует.");
     }
